requirements.txt
requests
cryptography
python_dotenv